refeactored to fix merge errors
