+++ App/controllers/auth.py
@@ -3,6 +3,16 @@ from flask_login import login_user, logout_user
 from App.models.user import User
 
 
+def authenticate(email, password):
+    user = User.query.filter_by(email=email).first()
+    if user and user.check_password(password):
+        return user
+
+# Payload is a dictionary which is passed to the function by Flask JWT
+def identity(payload):
+    return User.query.get(payload['identity'])
+
+
 def login():
     if request.method == 'GET':
         return render_template("login.html")
