--- App/database.py
@@ -1,4 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-
-# init db
-db = SQLAlchemy()
