+++ App/main.py
@@ -11,30 +11,23 @@ from flask_uploads import (
     TEXT,
     DOCUMENTS
 )
-from App.controllers.user import get_user_by_id
-from App.models.user import User
 
-from App.modules.auth_module import (authenticate, identity)
-from App.database import db
-
+from App.controllers import ( get_user_by_id, authenticate, identity)
+from App.models import (User, db)
 from App.views import (
-    api_bp,
-    user_bp,
-    chatroom_bp,
-    auth_bp,
-    api_views,
+    auth_views,
+    chat_views,
+    home_views,
+    topic_views,
     user_views,
-    user_registration
 )
 
-
 #place all views here
-views = [api_views, user_views , user_registration]
+views = [ auth_views, chat_views, home_views, topic_views, user_views]
 
 def add_views(app, views):
     for view in views:
         app.register_blueprint(view)
-
 
 def loadConfig(app, config):
     app.config['ENV'] = os.environ.get('ENV', 'development')
@@ -55,12 +48,10 @@ def init_db(app):
     db.init_app(app)
     db.create_all(app=app)
 
-
 def create_login_manager(app):
     login_manager = LoginManager()
     login_manager.init_app(app)
     return login_manager
-
 
 def create_app(config={}):
     app = Flask(__name__, static_url_path='/static')
