+++ App/views/__init__.py
@@ -1,4 +1,6 @@
-from .api_view import *
-from .user_view import *
-from .chatroom_view import *
-from .auth_view import *
+from .home import *
+from .user import *
+from .chat import *
+from .auth import *
+from .home import *
+from .topic import *
