--- App/views/api_view.py
@@ -1,7 +1,0 @@
-from flask import Blueprint, render_template
-
-api_bp = Blueprint('api_views', __name__, template_folder='../templates')
-
-@api_bp.route('/', methods=['GET'])
-def get_api_docs():
-    return render_template('index.html')
