--- App/views/auth_view.py
@@ -1,6 +1,0 @@
-from flask import Blueprint
-from App.controllers.auth import login, logout
-
-auth_bp = Blueprint("auth", __name__, template_folder='../templates')
-auth_bp.route("/login", methods=["GET", "POST"])(login)
-auth_bp.route("/logout", methods=["GET"])(logout)
